# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger Petstore",
        "description": "This is example for front end",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "yurayur1990@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://vatanparvarbackend.napaautomotive.uz/",
            "description": "This is domen for frontend"
        }
    ],
    "paths": {
...
        "/api/trusts/page/{count}": {
            "get": {
                "tags": [
                    "Board of Trustees"
                ],
                "summary": "Get data by short name",
                "operationId": "getTrustPag",
                "parameters": [
                    {
                        "name": "count",
                        "in": "path",
                        "description": "flage of country that needs to be got",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/api/sliders/{id}": {
            "get": {
                "tags": [
                    "Slider"
                ],
                "summary": "Get data by short name",
                "operationId": "getOneSlider",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "flage of country that needs to be got",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/api/infographics/{id}": {
            "get": {
                "tags": [
                    "Infographic"
                ],
                "summary": "Get data by short name",
                "operationId": "getOneInfographics",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "flage of country that needs to be got",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/infographics/page/{count}": {
            "get": {
                "tags": [
                    "Infographic"
                ],
                "summary": "Get data by short name",
                "operationId": "getInfographicsPag",
                "parameters": [
                    {
                        "name": "count",
                        "in": "path",
                        "description": "flage of country that needs to be got",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/api/events/{id}": {
            "get": {
                "tags": [
                    "Event"
                ],
                "summary": "Get data by short name",
                "operationId": "getOneEvents",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "flage of country that needs to be got",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/events/page/{count}": {
            "get": {
                "tags": [
                    "Event"
                ],
                "summary": "Get data by short name",
                "operationId": "getEventsPag",
                "parameters": [
                    {
                        "name": "count",
                        "in": "path",
                        "description": "flage of country that needs to be got",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/api/categoryshows/{id}": {
            "get": {
                "tags": [
                    "Turkum ko'rsatuvlar"
                ],
                "summary": "Get data by short name",
                "operationId": "getOneCategoryshows",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "flage of country that needs to be got",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/categoryshows/page/{count}": {
            "get": {
                "tags": [
                    "Turkum ko'rsatuvlar"
                ],
                "summary": "Get data by short name",
                "operationId": "getCategoryshowsPag",
                "parameters": [
                    {
                        "name": "count",
                        "in": "path",
                        "description": "flage of country that needs to be got",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
        "/api/country_relationship/{flage}": {
            "get": {
                "tags": [
                    "Country"
                ],
                "summary": "Get data by short name",
                "operationId": "countries",
                "parameters": [
                    {
                        "name": "flage",
                        "in": "path",
                        "description": "flage of country that needs to be got",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "string"
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                }
            }
        },
...
    }
}